Redirect home when no id given instead of 404ing

+++ module/Application/src/Application/Controller/TaskUpdateController.php
@@ -12,9 +12,14 @@ class TaskUpdateController extends AbstractActionController
     {
         $id = $this->params()->fromRoute('id');
 
+        // redirect back to home if not giving an id
+        if (!$id) return $this->redirect()->toRoute('home');
+
+        // grab the correct todo object
         $entityManager = $this->getServiceLocator()->get('Doctrine\ORM\EntityManager');
         $todo = $entityManager->getReference('Application\Entity\Todo', $id);
 
+        // toggle complete status
         if ($todo->isDone()) {
             $todo->markIncomplete();
         } else {
@@ -30,12 +35,16 @@ class TaskUpdateController extends AbstractActionController
     {
         $id = $this->params()->fromRoute('id');
 
+        // redirect back to home if not giving an id
+        if (!$id) return $this->redirect()->toRoute('home');
+
+        // grab the correct todo object
         $entityManager = $this->getServiceLocator()->get('Doctrine\ORM\EntityManager');
         $todo = $entityManager->getReference('Application\Entity\Todo', $id);
 
         $entityManager->remove($todo);
         $entityManager->flush();
 
-        return $this->redirect()->toRoute('home');        
+        return $this->redirect()->toRoute('home');
     }
 }
